+++ src/control/ModuleSettings.js
@@ -5,10 +5,13 @@ import {
    ViewManager }     from './index.js';
 
 import {
-   constants,
+   constants, keybindings,
    questStatus,
    sessionConstants,
-   settings }        from '../model/constants.js';
+   settings
+} from '../model/constants.js';
+
+import { QuestAPI } from "./public/index.js";
 
 /**
  * Provides registration for all module settings.
@@ -41,11 +44,20 @@ export class ModuleSettings
    };
 
    /**
+    * Registers all module settings and keybindings.
+    */
+   static register()
+   {
+      this.registerSettings();
+      this.registerKeybindings();
+   }
+
+   /**
     * Registers all module settings.
     *
     * @see FQLSettings
     */
-   static register()
+   static registerSettings()
    {
       game.settings.register(constants.moduleName, settings.allowPlayersDrag, {
          name: 'ForienQuestLog.Settings.allowPlayersDrag.Enable',
@@ -428,6 +440,65 @@ export class ModuleSettings
 
             // Swap macro image based on current state. No need to await.
             await Utils.setMacroImage(settings.questTrackerResizable, value);
+         }
+      });
+   }
+
+   /**
+    * Registers all module settings.
+    *
+    * @see FQLKeybindings
+    */
+   static registerKeybindings()
+   {
+
+      const { SHIFT, CONTROL, ALT } = foundry.helpers.interaction.KeyboardManager.MODIFIER_KEYS;
+
+      game.keybindings.register(constants.moduleName, keybindings.openQuestLog, {
+         name: `ForienQuestLog.Keybindings.openQuestLog.Name`,
+         hint: `ForienQuestLog.Keybindings.openQuestLog.Hint`,
+         editable: [
+            {
+               modifiers: [CONTROL],
+               key: "KeyQ",
+            }
+         ],
+         onDown: () =>
+         {
+            ViewManager.questLog.render(true);
+         }
+      });
+
+      game.keybindings.register(constants.moduleName, keybindings.openPrimaryQuest, {
+         name: `ForienQuestLog.Keybindings.openPrimaryQuest.Name`,
+         hint: `ForienQuestLog.Keybindings.openPrimaryQuest.Hint`,
+         precedence: CONST.KEYBINDING_PRECEDENCE.DEFERRED,
+         editable: [
+            {
+               modifiers: [SHIFT, CONTROL],
+               key: "KeyQ",
+            }
+         ],
+         onDown: () =>
+         {
+            QuestAPI.openPrimary();
+         }
+      });
+
+      game.keybindings.register(constants.moduleName, keybindings.toggleQuestTracker, {
+         name: `ForienQuestLog.Keybindings.toggleQuestTracker.Name`,
+         hint: `ForienQuestLog.Keybindings.toggleQuestTracker.Hint`,
+         precedence: CONST.KEYBINDING_PRECEDENCE.DEFERRED,
+         editable: [
+            {
+               modifiers: [SHIFT, ALT],
+               key: "KeyQ",
+            }
+         ],
+         onDown: () =>
+         {
+            const state = game.settings.get(constants.moduleName, settings.questTrackerEnable);
+            game.settings.set(constants.moduleName, settings.questTrackerEnable, !state);
          }
       });
    }